requirements.txt
pwntools
ropgadget